cli: move `wallet vote` to `validator vote`
